route calculation

+++ src/main/java/com/hackaton/task/trip/application/in/command/MakeRoadCommand.java
@@ -1,11 +1,10 @@
 package com.hackaton.task.trip.application.in.command;
 
 import com.hackaton.task.trip.domain.Attraction;
-import lombok.Builder;
 
 import java.util.List;
 
 
-public record MakeRoadCommand(Long distance,
+public record MakeRoadCommand(Integer distance,
                               List<Attraction> attractions) {
 }
